# src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Pagamentos",
    "version": "1.0.0",
    "description": "API RESTful para processamento de pagamentos com suporte a HATEOAS e JSON:API. Esta API permite criar, gerenciar e processar pagamentos com filtros avançados e paginação.",
    "contact": {
      "name": "Suporte",
      "email": "suporte@exemplo.com"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/payments/": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Cria um novo pagamento",
        "description": "Cria um novo pagamento no sistema com items e informações do pagador",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Dados do pagamento",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreatePaymentRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pagamento criado com sucesso",
            "schema": {
              "$ref": "#/definitions/PaymentResponse"
            }
          },
          "400": {
            "description": "Dados inválidos ou incompletos",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Lista todos os pagamentos com paginação",
        "description": "Obtém uma lista paginada de todos os pagamentos do sistema",
        "parameters": [
          {
            "name": "if-none-match",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página para paginação",
            "required": false,
            "type": "integer",
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Quantidade de itens por página",
            "required": false,
            "type": "integer",
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pagamentos com paginação",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "boolean"
                        },
                        "example": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "payments": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "array"
                                },
                                "items": {
                                  "$ref": "#/definitions/PaymentResponse"
                                }
                              }
                            },
                            "_meta": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "object"
                                },
                                "properties": {
                                  "type": "object",
                                  "properties": {
                                    "page": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "integer"
                                        },
                                        "example": {
                                          "type": "number",
                                          "example": 1
                                        }
                                      }
                                    },
                                    "limit": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "integer"
                                        },
                                        "example": {
                                          "type": "number",
                                          "example": 10
                                        }
                                      }
                                    },
                                    "total": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "integer"
                                        },
                                        "example": {
                                          "type": "number",
                                          "example": 100
                                        }
                                      }
                                    },
                                    "totalPages": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "integer"
                                        },
                                        "example": {
                                          "type": "number",
                                          "example": 10
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "_links": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "rel": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "next"
                                    }
                                  }
                                },
                                "href": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "/api/payments?page=2&limit=10"
                                    }
                                  }
                                },
                                "method": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "GET"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "304": {
            "description": "Not Modified"
          },
          "400": {
            "description": "Bad Request"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreatePaymentRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "required": {
          "type": "array",
          "example": [
            "userId",
            "orderId",
            "items",
            "payer"
          ],
          "items": {
            "type": "string"
          }
        },
        "properties": {
          "type": "object",
          "properties": {
            "userId": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "ID único do usuário"
                },
                "example": {
                  "type": "string",
                  "example": "user123"
                }
              }
            },
            "orderId": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "ID único do pedido"
                },
                "example": {
                  "type": "string",
                  "example": "order456"
                }
              }
            },
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "description": {
                  "type": "string",
                  "example": "Lista de itens do pagamento"
                },
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "required": {
                      "type": "array",
                      "example": [
                        "title",
                        "quantity",
                        "unitPrice"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "Nome do produto"
                            },
                            "example": {
                              "type": "string",
                              "example": "Smartphone Samsung Galaxy"
                            }
                          }
                        },
                        "description": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "Descrição detalhada do produto"
                            },
                            "example": {
                              "type": "string",
                              "example": "Smartphone Samsung Galaxy A54 128GB Preto"
                            }
                          }
                        },
                        "quantity": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "integer"
                            },
                            "minimum": {
                              "type": "number",
                              "example": 1
                            },
                            "description": {
                              "type": "string",
                              "example": "Quantidade do item"
                            },
                            "example": {
                              "type": "number",
                              "example": 2
                            }
                          }
                        },
                        "unitPrice": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "number"
                            },
                            "minimum": {
                              "type": "number",
                              "example": 0.01
                            },
                            "description": {
                              "type": "string",
                              "example": "Preço unitário em reais"
                            },
                            "example": {
                              "type": "number",
                              "example": 899.99
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "Fone de Ouvido Bluetooth"
                      },
                      "description": {
                        "type": "string",
                        "example": "Fone de Ouvido Bluetooth JBL Tune 760NC"
                      },
                      "quantity": {
                        "type": "number",
                        "example": 1
                      },
                      "unitPrice": {
                        "type": "number",
                        "example": 299.99
                      }
                    }
                  }
                }
              }
            },
            "payer": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "required": {
                  "type": "array",
                  "example": [
                    "email",
                    "name",
                    "identification"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "email"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email do pagador"
                        },
                        "example": {
                          "type": "string",
                          "example": "joao.silva@email.com"
                        }
                      }
                    },
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Nome completo do pagador"
                        },
                        "example": {
                          "type": "string",
                          "example": "João Silva Santos"
                        }
                      }
                    },
                    "identification": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "required": {
                          "type": "array",
                          "example": [
                            "type",
                            "number"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "enum": {
                                  "type": "array",
                                  "example": [
                                    "CPF",
                                    "CNPJ"
                                  ],
                                  "items": {
                                    "type": "string"
                                  }
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Tipo de documento"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "CPF"
                                }
                              }
                            },
                            "number": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Número do documento"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "12345678901"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "callbackUrl": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uri"
                },
                "description": {
                  "type": "string",
                  "example": "URL para callback após processamento do pagamento"
                },
                "example": {
                  "type": "string",
                  "example": "https://meusite.com/webhook/payment"
                }
              }
            },
            "authToken": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Token de autenticação (opcional)"
                },
                "example": {
                  "type": "string",
                  "example": "auth_token_123"
                }
              }
            }
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "success": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "boolean"
                },
                "example": {
                  "type": "boolean",
                  "example": false
                }
              }
            },
            "message": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Dados incompletos para criação do pagamento"
                }
              }
            },
            "error": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "userId, orderId, items e payer são obrigatórios"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  }
}